separate glb format into separate file
